Remove stub
