+++ bika/health/upgrade/v01_02_000.py
@@ -25,11 +25,11 @@ from bika.health.setuphandlers import remove_action
 from bika.health.setuphandlers import setup_content_actions
 from bika.health.setuphandlers import setup_id_formatting
 from bika.health.setuphandlers import setup_roles_permissions
-from bika.health.subscribers.patient import purge_owners_for
 from bika.health.upgrade.utils import del_column
 from bika.health.upgrade.utils import del_index
 from bika.health.upgrade.utils import setup_catalogs
 from bika.lims import api
+from bika.lims.api import security
 from bika.lims.upgrade import upgradestep
 from bika.lims.upgrade.utils import UpgradeUtils
 from bika.lims.upgrade.utils import commit_transaction
@@ -184,3 +184,45 @@ def remove_stale_css(portal):
     for css in CSS_TO_REMOVE:
         logger.info("Unregistering CSS %s" % css)
         portal.portal_css.unregisterResource(css)
+
+
+# TODO: This is no longer needed!
+def assign_owners_for(patient):
+    """Assign the role "Owner" to the contacts of the client assigned to the
+    patient passed in, if any
+    """
+    client = patient.getClient()
+    if not client:
+        return False
+
+    contacts = client.objectValues("Contact")
+    users = map(lambda contact: contact.getUser(), contacts)
+    users = filter(None, users)
+    for user in users:
+        security.grant_local_roles_for(patient, roles=["Owner"], user=user)
+    patient.reindexObjectSecurity()
+    return True
+
+
+def purge_owners_for(patient):
+    """Remove role "Owner" from all those client contacts that do not belong to
+    the same Client the patient is assigned to and assigns the role "Owner" to
+    the client contacts assigned to the patient
+    """
+    # Add role "Owner" for this Patient to all contacts from this Client
+    assign_owners_for(patient)
+
+    # Unassign role "Owner" from contacts that belong to another Client
+    patient_client = patient.getClient()
+    patient_client_uid = patient_client and api.get_uid(patient_client) or None
+    for client in api.search(dict(portal_type="Client"), "portal_catalog"):
+        if api.get_uid(client) == patient_client_uid:
+            continue
+
+        client = api.get_object(client)
+        contacts = client.objectValues("Contact")
+        users = map(lambda contact: contact.getUser(), contacts)
+        users = filter(None, users)
+        for user in users:
+            security.revoke_local_roles_for(patient, ["Owner"], user=user)
+    patient.reindexObjectSecurity()
